requirements.txt
fastapi==0.115.12
numpy>=1.26.0,<2.2.0
Pillow==11.2.1
pydantic==2.11.5
tensorflow==2.19.0
uvicorn==0.34.2
python-multipart==0.0.9